dispositivo/serializers.py
from rest_framework import serializers
from .models import Dispositivo

class DispositivoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Dispositivo
        fields = '__all__'
