requirements.txt
Django==5.0.3
django-colorfield==0.3.2
django-cors-headers==3.4.0
django-filter==23.5
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0

